Moved ReposCache to its own package. Moved ReposCallback to its own package.
