+++ app/src/main/java/com/andevcba/githubmvp/data/ReposCallback.java
@@ -1,4 +1,4 @@
-package com.andevcba.githubmvp.data.repository;
+package com.andevcba.githubmvp.data;
 
 import com.andevcba.githubmvp.data.model.ReposByUsername;
 
@@ -8,6 +8,7 @@ import com.andevcba.githubmvp.data.model.ReposByUsername;
  * @author lucas.nobile
  */
 public interface ReposCallback {
+
     void onResponse(ReposByUsername response);
 
     void onError(String error);
