+++ app/src/main/java/com/andevcba/githubmvp/data/repository/InMemoryRepository.java
@@ -16,6 +16,8 @@
 
 package com.andevcba.githubmvp.data.repository;
 
+import com.andevcba.githubmvp.data.ReposCallback;
+import com.andevcba.githubmvp.data.cache.ReposCache;
 import com.andevcba.githubmvp.data.model.Repo;
 import com.andevcba.githubmvp.data.model.ReposByUsername;
