+++ app/src/main/java/com/andevcba/githubmvp/domain/interactor/LoadReposInteractor.java
@@ -1,8 +1,8 @@
 package com.andevcba.githubmvp.domain.interactor;
 
 import com.andevcba.githubmvp.data.DependencyProvider;
-import com.andevcba.githubmvp.data.repository.ReposCache;
-import com.andevcba.githubmvp.data.repository.ReposCallback;
+import com.andevcba.githubmvp.data.ReposCallback;
+import com.andevcba.githubmvp.data.cache.ReposCache;
 import com.andevcba.githubmvp.data.repository.Repository;
 
 /**
