+++ app/src/main/java/com/andevcba/githubmvp/domain/interactor/SaveReposInteractor.java
@@ -1,8 +1,8 @@
 package com.andevcba.githubmvp.domain.interactor;
 
 import com.andevcba.githubmvp.data.DependencyProvider;
+import com.andevcba.githubmvp.data.cache.ReposCache;
 import com.andevcba.githubmvp.data.model.ReposByUsername;
-import com.andevcba.githubmvp.data.repository.ReposCache;
 import com.andevcba.githubmvp.data.repository.Repository;
 
 /**
