+++ app/src/test/java/com/andevcba/githubmvp/domain/interactor/SaveReposInteractorTest.java
@@ -1,10 +1,10 @@
 package com.andevcba.githubmvp.domain.interactor;
 
+import com.andevcba.githubmvp.data.ReposCallback;
+import com.andevcba.githubmvp.data.cache.ReposCache;
 import com.andevcba.githubmvp.data.model.Repo;
 import com.andevcba.githubmvp.data.model.ReposByUsername;
 import com.andevcba.githubmvp.data.repository.InMemoryRepository;
-import com.andevcba.githubmvp.data.repository.ReposCache;
-import com.andevcba.githubmvp.data.repository.ReposCallback;
 
 import org.junit.Before;
 import org.junit.Test;
